Create README.md
The book-selling website offers a diverse library of rich titles, covering genres from literature, fiction, economics to science and art. Each book is described in detail, with slides and information about the author and price.

Users can easily search for books by name, author or by genre using the smart search bar. Once the desired book is found, the user can view detailed information and choose to add it to the cart. The buying process is simple and secure, with convenient and secure online payment methods.